MainPage: Fix the Flickable behaviour
We don't want the flickable bouncy behaviour when overshooting. Also,
the playhead will be centered in the middle of the timeline and fixed in
that position. Thus initially the text must begin at the centre of the
timeline area and sliding the text all the way to the left should leave
the end in the middle also.