Timeline list item appending stuff
Create a TimelineList ListModel that is instantiated in main to be
shared between pages. Add code to the video gallery page to append items
to the timelineListModel when the 'Add' button is pressed. Create a
ListView in the timelineBar to display the list of items added.
Currently just the URI is displayed as text.