videoeditor: adds a new qml component VideoEditor
VideoEditor will be the placeholder for the GES elements. It should
be responsible for managing the list of tracks and medias and be able
to render and play them on the screen.

It is also a ListModel, exposing the medias it currently has, so that
they can be displayed easily on the UI.

For now it is merely a dumb listmodel that holds the URLs of the media
it should concatenate