Timeline: Add export file dialog
When the export button is pressed, the user is presented with a text
input box to input a filename. The filename is populated by default with
a date-time-count.mp4 name which can be edited to be whatever. The file
will always be saved in /home/user/MyDocs/Movies and will appear in the
Videos app. The filename is maintained within the session so, for
example, if one edits the filename, cancels and then hits export again,
the filename will be as edited.