Timeline: Drag and drop only selected item
Make some alterations to the ListView delegate to allow us to select an
item and only drag and drop the selected item. There's a bug with the
selected item when dropping, I suspect due to the list reset when
dropping and moving. Perhaps the index can be restored across this
operation.