videoeditor: Work towards exposing object and timeline durations
This commit adds a VideoEditorItem class and a list of pointers to such
objects such that we can manage a list internally in the
QDeclarativeVideoEditor rather than having to look up all properties
from GES with potentially expensive queries.

The object duration is obtained from the GESTimelineFileSource
max-duration property. When discovery for an object completes, as long
as inPoint is 0, max-duration will be the duration of the clip. We need
to make sure to store the clip duration separately from the timeline
object duration when we get to trimming.

The timeline duration should be the longer of the audio and video track
durations. In newer GES versions the timeline has its own duration
property but we will have to work around that omission due to the age of
the GES version included in Harmattan. This is not currently working
however and this needs more work.