capstricks: adds more caps manipulation tricks
In a line: make some elements getcaps/setcaps work as needed

Long version:
* dsp and aac encoder elements don't proxy downstream caps restrictions
to upstream, so that the converters can act on framerate/resolution
rate/channels and avoid negotiation failures when it is too late.
We add our own getcaps function to those elements to make sure they
behave correctly.

* When setting a new caps after finishing one media and starting another one,
the dsp encoder might not put the codec_data at the caps at first, but it adds
it later. This won't work as mp4mux will reject a new caps that doesn't match
the previous one. The solution was to add a new setcaps that makes sure the
caps is the same and passes it to the old setcaps function of mp4mux pads.