videoeditor: avoid using smart-render mode
The smart-render mode will try to avoid decoding a stream to reencode
to the same format. This would reduce the cpu usage, but it also increases
the change of negotiation failures when concatenating videos of almost the
same format. GES might interpret them of being the same format, but the
muxer will reject a caps change, even if the differences are minimal.

This makes the app more robust