timeline: workaround for listview stubbornness with contentX
When a new item is selected, listview attempts to position it at
the visible area, if it is larger then the screen, it will position
its beginning aligned with the left of the timeline, making it impossible
for the user to use the controls at the end of the clip.

This hacky patch works around this. The QML issue was reported at:
https://bugreports.qt-project.org/browse/QTBUG-25154

Fixes VideoEditor issue #19