timeline: always move the contentX to 0 before scaling
Doing the scaling when contentX isn't 0 messes up the timeline items
placement, which afects any of the features that depends on the position
of the the playhead (playback, seeking, trimming).

This is not very pleasant to the user, but it is better than
having bad results for the final rendering.

Fixes issue #28