videoeditor: check if we are on PR1.3 or newer to enable thumbnails
PR1.2 has a bug that makes image loading crash. Check if we are
running PR1.3 or newer before enabling thumbnails.

To check, the only option found was to look at the about applet
binary contents and look for PRX.Y string, it seems this value
isn't found in any other system's file.