README: Update build instructions
It seems the libges libs linked to did not contain the following
required API:

ges_timeline_pipeline_preview_set_video_sink ()

Therefore an updated ges-timeline-pipeline.h and libges.so.0.0.0 are
added in the dependencies directory. This binary shared object is copied
from an N9 with publicly available PR1.3. Nothing special, anyone with
the phone could get it off there.